rename types to model
