+++ internal/config/db.go
@@ -3,7 +3,7 @@ package config
 import (
 	"log"
 
-	"github.com/b-open/jobbuzz/pkg/types"
+	"github.com/b-open/jobbuzz/pkg/model"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
@@ -20,5 +20,5 @@ func GetDb() *gorm.DB {
 }
 
 func MigrateDb(db *gorm.DB) {
-	db.AutoMigrate(&types.Job{})
+	db.AutoMigrate(&model.Job{})
 }
