+++ pkg/model/model.go
@@ -1,4 +1,4 @@
-package types
+package model
 
 import "time"
